add additional fuzzing coverage for various APIs
* refactor, move Creation and Action into their own modules
* fuzz from_utf8_buf_unchecked
* fuzz with_capacity(usize)
* fuzz reserve(usize)
* fuzz FromIterator<char>, also adds coverage for larger strings
* fuzz the Default impl
* fuzz as_mut_bytes
* fuzz fmt::Debug impl
* fuzz fmt::Display impl
* fuzz From<Cow<str>>
* fuzz FromIterator<String>
* fuzz FromIterator<Box<str>>

* pin version of Rust nightly when fuzzing with AFL on ARMv7
